@@ -9,29 +9,29 @@
     "lint": "eslint './**/*.ts{,x}'"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "^13.1.6",
-    "firebase": "^9.17.1",
-    "next": "^13.1.6",
+    "@next/bundle-analyzer": "^13.2.4",
+    "firebase": "^9.18.0",
+    "next": "^13.2.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
-    "react-spring": "^9.6.1",
-    "styled-components": "^5.3.6"
+    "react-spring": "^9.7.1",
+    "styled-components": "^5.3.9"
   },
   "devDependencies": {
-    "@types/node": "^18.14.0",
-    "@types/react": "^18.0.28",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.31",
     "@types/react-modal": "^3.13.1",
     "@types/styled-components": "^5.1.26",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "@typescript-eslint/parser": "^5.52.0",
-    "eslint": "^8.34.0",
-    "eslint-config-next": "^13.1.6",
-    "eslint-config-prettier": "^8.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "eslint": "^8.37.0",
+    "eslint-config-next": "^13.2.4",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.8.4",
-    "typescript": "^4.9.5"
+    "prettier": "^2.8.7",
+    "typescript": "^5.0.2"
   }
 }
